+++ smoke-tests/apps/PreAggMetricsWithRoleNameOverridesAndSampling/src/main/java/com/microsoft/applicationinsights/smoketestapp/PreAggMetricsWithRoleNameOverridesAndSamplingServlet.java
@@ -28,7 +28,7 @@ public class PreAggMetricsWithRoleNameOverridesAndSamplingServlet extends HttpSe
       return;
     }
     if (pathInfo.startsWith("/app")) {
-      httpUrlConnection("https://mock.codes/200");
+      httpUrlConnection("http://host.testcontainers.internal:6060/mock/200");
       logger.info("hello");
       return;
     }